etcd: rename pkg/{etcd -> embeddedetcd}
