+++ pkg/server/options/options.go
@@ -35,14 +35,14 @@ import (
 	kubeoptions "k8s.io/kubernetes/pkg/kubeapiserver/options"
 
 	kcpadmission "github.com/kcp-dev/kcp/pkg/admission"
-	etcdoptions "github.com/kcp-dev/kcp/pkg/etcd/options"
+	etcdoptions "github.com/kcp-dev/kcp/pkg/embeddedetcd/options"
 	_ "github.com/kcp-dev/kcp/pkg/features"
 	kcpfeatures "github.com/kcp-dev/kcp/pkg/features"
 )
 
 type Options struct {
 	GenericControlPlane ServerRunOptions
-	EmbeddedEtcd        etcdoptions.Embedded
+	EmbeddedEtcd        etcdoptions.Options
 	Controllers         Controllers
 	Authorization       Authorization
 	AdminAuthentication AdminAuthentication
@@ -66,7 +66,7 @@ type ExtraOptions struct {
 
 type completedOptions struct {
 	GenericControlPlane options.CompletedServerRunOptions
-	EmbeddedEtcd        etcdoptions.CompletedEmbedded
+	EmbeddedEtcd        etcdoptions.CompletedObjects
 	Controllers         Controllers
 	Authorization       Authorization
 	AdminAuthentication AdminAuthentication
@@ -86,7 +86,7 @@ func NewOptions(rootDir string) *Options {
 		GenericControlPlane: ServerRunOptions{
 			*options.NewServerRunOptions(),
 		},
-		EmbeddedEtcd:        *etcdoptions.NewEmbedded(rootDir),
+		EmbeddedEtcd:        *etcdoptions.NewOptions(rootDir),
 		Controllers:         *NewControllers(),
 		Authorization:       *NewAuthorization(),
 		AdminAuthentication: *NewAdminAuthentication(rootDir),
@@ -193,7 +193,7 @@ func (o *CompletedOptions) Validate() []error {
 }
 
 func (o *Options) Complete() (*CompletedOptions, error) {
-	var completedEmbeddedEtcd etcdoptions.CompletedEmbedded
+	var completedEmbeddedEtcd etcdoptions.CompletedObjects
 	if servers := o.GenericControlPlane.Etcd.StorageConfig.Transport.ServerList; len(servers) == 1 && servers[0] == "embedded" {
 		o.GenericControlPlane.Etcd.StorageConfig.Transport.ServerList = []string{"localhost:" + o.EmbeddedEtcd.ClientPort}
 		o.EmbeddedEtcd.Enabled = true
